# backend/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Hambin API
  version: 1.0.0
  description: >
    Client-neutral HTTP API for room management, presence, shared stream
    authorization, realtime tickets, and LiveKit voice tokens. Clients retain
    complete local control over pause, seek, and volume; those actions never
    mutate durable stream state.
servers:
  - url: https://api.example.com
    description: Production
...
paths:
...
  /v1/guest-sessions/profile:
    patch:
      tags: [Guest sessions]
      operationId: updateProfile
      summary: Update the caller's display name, avatar, and language
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GuestSessionRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                type: object
                required: [identity]
                properties:
                  identity:
                    $ref: "#/components/schemas/GuestIdentity"
        "400":
          $ref: "#/components/responses/BadRequest"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    GuestSessionRequest:
      type: object
      required: [name, avatar, language]
      properties:
        name:
          type: string
          maxLength: 48
        avatar:
          type: string
          maxLength: 32
        language:
          type: string
          maxLength: 16
...
    GuestIdentity:
      type: object
      required: [id, name, avatar, language, createdAt, updatedAt]
      properties:
        id:
          type: string
        name:
          type: string
        avatar:
          type: string
        language:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum:
            - INVALID_JSON
            - INVALID_PROFILE
            - INVALID_TOKEN
            - UNAUTHENTICATED
            - SESSION_EXPIRED
            - SESSION_REVOKED
            - IDENTITY_NOT_FOUND
            - ROOM_NOT_FOUND
            - ROOM_ENDED
            - ROOM_FULL
            - ALREADY_IN_ROOM
            - NOT_A_MEMBER
            - INVALID_PARTY_CODE
            - FORBIDDEN
            - OWNER_ABSENT
            - INVALID_STREAM_URL
            - STALE_STREAM_REVISION
            - NO_ACTIVE_STREAM
            - INVALID_TICKET
            - VOICE_DISABLED
            - INVALID_WEBHOOK
            - RATE_LIMITED
            - INTERNAL_ERROR
        message:
          type: string